requirements.txt
Django==6.0.4
groq==1.1.2
python-dotenv==1.2.2
